refactor(routes): extract jellyfin command routes
